+++ quilt-cli/src/cli/model.rs
@@ -8,6 +8,7 @@ use quilt_rs::lineage::Home;
 
 use crate::cli::browse;
 use crate::cli::commit;
+use crate::cli::create;
 use crate::cli::install;
 use crate::cli::list;
 use crate::cli::login;
@@ -27,6 +28,11 @@ pub trait Commands {
     async fn browse(&self, args: browse::Input) -> Result<browse::Output, Error> {
         let local_domain = self.get_local_domain();
         browse::model(local_domain, args).await
+    }
+
+    async fn create(&self, args: create::Input) -> Result<create::Output, Error> {
+        let local_domain = self.get_local_domain();
+        create::model(local_domain, args).await
     }
 
     async fn commit(&self, args: commit::Input) -> Result<commit::Output, Error> {