+++ quilt-rs/src/flow.rs
@@ -6,6 +6,7 @@
 mod browse;
 mod certify_latest;
 mod commit;
+mod create_package;
 mod install_package;
 mod install_paths;
 mod pull;
@@ -19,6 +20,7 @@ pub use browse::browse_remote_manifest as browse;
 pub use browse::cache_remote_manifest;
 pub use certify_latest::certify_latest;
 pub use commit::commit_package as commit;
+pub use create_package::create_package as create;
 pub use install_package::install_package;
 pub use install_paths::install_paths;
 pub use pull::pull_package as pull;